Changed Delete Logic
Looked at the assignment doc and made changes to delete to fit the following. "After deletion, the next song in the list should be selected, and the details displayed. If there is no next song, the previous song should be selected, and if there is no previous song either, then the list is empty and the detail info is all blanks."